Solves problem 3: Largest Prime Factor
A little harder, given it's been years since I studied maths.

At first I thought about generating a list of primes and checking against that. I realised that would be very expensive, instead, I check if two is a factor, if not, I check if any odd numbers are factors.

This feels like a brute force method, I'm sure it could be improved.